server/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.5
exceptiongroup==1.3.1
fastapi==0.128.8
h11==0.16.0
idna==3.11
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==11.3.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pypdfium2==5.5.0
python-multipart==0.0.20
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.39.0
lxml
python-multipart
python-docx
beautifulsoup4
